requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==38.0.4
daphne==4.0.0
distlib==0.3.6
Django==4.1.3
django-ckeditor==6.5.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.7
Faker==15.3.3
filelock==3.8.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
mysqlclient==2.1.1
Pillow==9.3.0
platformdirs==2.5.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-dateutil==2.8.2
python-decouple==3.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.6
validate-email==1.3
virtualenv==20.16.7
zope.interface==5.5.2
